# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is a sample of RESTful api made by cargo-rest-api",
        "title": "cargo-rest-api",
        "contact": {
            "name": "Aleksand Efremov",
            "url": "https://github.com/efremovich/cargo-rest-api",
            "email": "sasha.fima@gmail.com"
        },
        "license": {
            "name": "MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8888",
    "paths": {
...
        "/api/v1/external/preference": {
            "get": {
                "security": [
                    {
                        "BasicAuth": []
                    },
                    {
                        "JWTAuth": []
                    }
                ],
                "description": "Get current user preference using Authorization Header.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "preference"
                ],
                "summary": "User preference",
                "parameters": [
                    {
                        "enum": [
                            "en",
                            "ru"
                        ],
                        "type": "string",
                        "default": "en",
                        "description": "Language code",
                        "name": "Accept-Language",
                        "in": "header"
                    },
                    {
                        "type": "string",
                        "description": "Request id",
                        "name": "Set-Request-Id",
                        "in": "header"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.successOutput"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.errorOutput"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.errorOutput"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.errorOutput"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.errorOutput"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BasicAuth": []
                    },
                    {
                        "JWTAuth": []
                    }
                ],
                "description": "Update current user preference using Authorization Header.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "preference"
                ],
                "summary": "Update user preference",
                "parameters": [
                    {
                        "enum": [
                            "en",
                            "ru"
                        ],
                        "type": "string",
                        "default": "en",
                        "description": "Language code",
                        "name": "Accept-Language",
                        "in": "header"
                    },
                    {
                        "type": "string",
                        "description": "Request id",
                        "name": "Set-Request-Id",
                        "in": "header"
                    },
                    {
                        "description": "User preference",
                        "name": "preference",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.DetailUserPreference"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.successOutput"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.errorOutput"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.errorOutput"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.errorOutput"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/response.errorOutput"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.errorOutput"
                        }
                    }
                }
            }
        },
        "/api/v1/external/preference/reset": {
            "post": {
                "security": [
                    {
                        "BasicAuth": []
                    },
                    {
                        "JWTAuth": []
                    }
                ],
                "description": "Reset current user preference to default using Authorization Header.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "preference"
                ],
                "summary": "Reset user preference",
                "parameters": [
                    {
                        "enum": [
                            "en",
                            "ru"
                        ],
                        "type": "string",
                        "default": "en",
                        "description": "Language code",
                        "name": "Accept-Language",
                        "in": "header"
                    },
                    {
                        "type": "string",
                        "description": "Request id",
                        "name": "Set-Request-Id",
                        "in": "header"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.successOutput"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.errorOutput"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.errorOutput"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.errorOutput"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.errorOutput"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "entity.DetailUserPreference": {
            "type": "object",
            "properties": {
                "dark_mode": {
                    "type": "boolean"
                },
                "language": {
                    "type": "string"
                },
                "skip_tour": {
                    "type": "boolean"
                }
            }
        },
        "response.errorOutput": {
            "type": "object",
            "properties": {
                "args": {
                    "type": "string"
                },
                "code": {
                    "type": "integer"
                },
                "data": {},
                "error_code": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "response.successOutput": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                },
                "meta": {}
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "api_key",
            "in": "header"
        },
        "BasicAuth": {
            "type": "basic"
        },
        "JWTAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "Oauth2AccessCode": {
            "type": "oauth2",
            "flow": "accessCode",
            "authorizationUrl": "http://localhost:8888/oauth/authorize",
            "tokenUrl": "http://localhost:8888/oauth/token",
            "scopes": {
                "all": " Grants all access"
            }
        },
        "Oauth2Password": {
            "type": "oauth2",
            "flow": "password",
            "tokenUrl": "http://localhost:8888/oauth/token",
            "scopes": {
                "all": " Grants all access"
            }
        }
    }
}